@@ -3,41 +3,43 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --host 0.0.0.0 ",
+    "start": "ng serve --proxy-config proxy.conf.json --host 0.0.0.0 ",
     "start-prod": "ng serve --host 0.0.0.0 --configuration=production",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
-    "test": "ng test"
+    "test": "ng test",
+    "cy": "cypress run"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.5",
-    "@angular/cdk": "^13.3.5",
-    "@angular/common": "~13.3.5",
-    "@angular/compiler": "~13.3.5",
-    "@angular/core": "~13.3.5",
-    "@angular/flex-layout": "^13.0.0-beta.38",
-    "@angular/forms": "~13.3.5",
-    "@angular/material": "^13.3.5",
-    "@angular/platform-browser": "~13.3.5",
-    "@angular/platform-browser-dynamic": "~13.3.5",
-    "@angular/router": "~13.3.5",
-    "rxjs": "~7.5.5",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "@angular/animations": "~15.2.0",
+    "@angular/cdk": "~15.2.9",
+    "@angular/common": "~15.2.10",
+    "@angular/compiler": "~15.2.10",
+    "@angular/core": "~15.2.10",
+    "@angular/flex-layout": "~15.0.0-beta.42",
+    "@angular/forms": "~15.2.10",
+    "@angular/material": "~15.2.9",
+    "@angular/platform-browser": "~15.2.10",
+    "@angular/platform-browser-dynamic": "~15.2.10",
+    "@angular/router": "~15.2.10",
+    "@think-it-labs/edc-connector-client": "^0.2.0-beta-7",
+    "rxjs": "~7.8.1",
+    "zone.js": "~0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.4",
-    "@angular/cli": "~13.3.4",
-    "@angular/compiler-cli": "~13.3.5",
+    "@angular-devkit/build-angular": "~15.2.10",
+    "@angular/cli": "~15.2.10",
+    "@angular/compiler-cli": "~15.2.9",
     "@types/jasmine": "~4.0.3",
-    "@types/node": "^17.0.30",
-    "jasmine-core": "~4.1.0",
+    "@types/node": "^20.8.7",
+    "cypress": "^13.3.2",
+    "jasmine-core": "~5.1.1",
     "karma": "~6.3.16",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
+    "karma-coverage": "~2.2.1",
     "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.6.4"
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "typescript": "~4.8.2"
   }
 }
