@@ -1,6 +1,5 @@
 import {Component, Inject, OnInit} from '@angular/core';
-import {Policy, PolicyDefinitionResponseDto, PolicyDefinitionRequestDto} from "../../../mgmt-api-client";
-import TypeEnum = Policy.TypeEnum;
+import {PolicyDefinitionInput, PolicyInput} from "../../../mgmt-api-client/model";
 import {MAT_DIALOG_DATA, MatDialogRef} from "@angular/material/dialog";
 
 @Component({
@@ -10,12 +9,12 @@ import {MAT_DIALOG_DATA, MatDialogRef} from "@angular/material/dialog";
 })
 export class NewPolicyDialogComponent implements OnInit {
   editMode: boolean = false;
-  policy: Policy = {
-    type: TypeEnum.Set
+  policy: PolicyInput = {
+    "@type": "set"
   };
-  policyDefinition: PolicyDefinitionResponseDto = {
-    policy: this.policy,
-    id: ''
+  policyDefinition: PolicyDefinitionInput = {
+    "policy": this.policy,
+    "@id": ''
   };
   permissionsJson: string = '';
   prohibitionsJson: string = '';
@@ -30,20 +29,24 @@ export class NewPolicyDialogComponent implements OnInit {
 
   onSave() {
     if (this.permissionsJson && this.permissionsJson !== '') {
-      this.policy.permissions = JSON.parse(this.permissionsJson);
+      this.policy.permission = JSON.parse(this.permissionsJson);
     }
 
     if (this.prohibitionsJson && this.prohibitionsJson !== '') {
-      this.policy.prohibitions = JSON.parse(this.prohibitionsJson);
+      this.policy.prohibition = JSON.parse(this.prohibitionsJson);
     }
 
     if (this.obligationsJson && this.obligationsJson !== '') {
-      this.policy.obligations = JSON.parse(this.obligationsJson);
+      this.policy.obligation = JSON.parse(this.obligationsJson);
     }
 
+    this.policy["@context"]="http://www.w3.org/ns/odrl.jsonld"
+
+
     this.dialogRef.close({
-      policy: this.policyDefinition.policy,
-      id: this.policyDefinition.id
+
+      policy : this.policyDefinition.policy,
+      '@id': this.policyDefinition.id
     })
   }
 }
