@@ -1,6 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {Observable, of} from 'rxjs';
-import {TransferProcessDto, TransferProcessService} from "../../../mgmt-api-client";
+import {TransferProcessService} from "../../../mgmt-api-client";
+import {TransferProcess} from "../../../mgmt-api-client/model";
 import {AppConfigService} from "../../../app/app-config.service";
 import {ConfirmationDialogComponent, ConfirmDialogModel} from "../confirmation-dialog/confirmation-dialog.component";
 import {MatDialog} from "@angular/material/dialog";
@@ -12,8 +13,8 @@ import {MatDialog} from "@angular/material/dialog";
 })
 export class TransferHistoryViewerComponent implements OnInit {
 
-  columns: string[] = ['id', 'creationDate', 'state', 'lastUpdated', 'connectorId', 'assetId', 'contractId', 'action'];
-  transferProcesses$: Observable<TransferProcessDto[]> = of([]);
+  columns: string[] = ['id', 'state', 'lastUpdated', 'connectorId', 'assetId', 'contractId', 'action'];
+  transferProcesses$: Observable<TransferProcess[]> = of([]);
   storageExplorerLinkTemplate: string | undefined;
 
   constructor(private transferProcessService: TransferProcessService,
@@ -26,9 +27,9 @@ export class TransferHistoryViewerComponent implements OnInit {
     this.storageExplorerLinkTemplate = this.appConfigService.getConfig()?.storageExplorerLinkTemplate
   }
 
-  onDeprovision(transferProcess: TransferProcessDto): void {
+  onDeprovision(transferProcess: TransferProcess): void {
 
-    const dialogData = new ConfirmDialogModel("Confirm deprovision", `Deprovisioning resources for transfer [${transferProcess.id}] will take some time and once started, it cannot be stopped.`)
+    const dialogData = new ConfirmDialogModel("Confirm deprovision", `Deprovisioning resources for transfer [${transferProcess["@id"]}] will take some time and once started, it cannot be stopped.`)
     dialogData.confirmColor = "warn";
     dialogData.confirmText = "Confirm";
     dialogData.cancelText = "Abort";
@@ -36,21 +37,21 @@ export class TransferHistoryViewerComponent implements OnInit {
 
     ref.afterClosed().subscribe(res => {
       if (res) {
-        this.transferProcessService.deprovisionTransferProcess(transferProcess.id!).subscribe(() => this.loadTransferProcesses());
+       this.transferProcessService.deprovisionTransferProcess(transferProcess["@id"]!).subscribe(() => this.loadTransferProcesses());
       }
     });
   }
 
-  showStorageExplorerLink(transferProcess: TransferProcessDto) {
+  showStorageExplorerLink(transferProcess: TransferProcess) {
     return transferProcess.dataDestination?.properties?.type === 'AzureStorage' && transferProcess.state === 'COMPLETED';
   }
 
-  showDeprovisionButton(transferProcess: TransferProcessDto) {
+  showDeprovisionButton(transferProcess: TransferProcess) {
     return ['COMPLETED', 'PROVISIONED', 'REQUESTED', 'REQUESTED_ACK', 'IN_PROGRESS', 'STREAMING'].includes(transferProcess.state!);
   }
 
   loadTransferProcesses() {
-    this.transferProcesses$ = this.transferProcessService.getAllTransferProcesses();
+     this.transferProcesses$ = this.transferProcessService.queryAllTransferProcesses();
   }
 
   asDate(epochMillis?: number) {
