@@ -1,10 +1,12 @@
-import { Asset } from "./asset";
-import {Policy} from "../../mgmt-api-client";
+import {DataService} from "./data-service";
+import {PolicyInput} from "../../mgmt-api-client/model";
 
 export interface ContractOffer {
-    id: string;
-    policy: Policy;
-    provider: string;
-    consumer: string;
-    asset: Asset;
+  id: string;
+  assetId: string;
+  properties: any;
+  "dcat:dataset": Array<any>;
+  "dcat:service": DataService;
+  policy: PolicyInput;
+  originator: string;
 }
