@@ -1,21 +1,13 @@
-export * from './applicationObservability.service';
-import { ApplicationObservabilityService } from './applicationObservability.service';
 export * from './asset.service';
 import { AssetService } from './asset.service';
-export * from './catalog.service';
-import { CatalogService } from './catalog.service';
 export * from './contractAgreement.service';
 import { ContractAgreementService } from './contractAgreement.service';
 export * from './contractDefinition.service';
 import { ContractDefinitionService } from './contractDefinition.service';
 export * from './contractNegotiation.service';
 import { ContractNegotiationService } from './contractNegotiation.service';
-export * from './dataplaneSelector.service';
-import { DataplaneSelectorService } from './dataplaneSelector.service';
-export * from './hTTPProvisionerWebhook.service';
-import { HTTPProvisionerWebhookService } from './hTTPProvisionerWebhook.service';
 export * from './policy.service';
 import { PolicyService } from './policy.service';
 export * from './transferProcess.service';
 import { TransferProcessService } from './transferProcess.service';
-export const APIS = [ApplicationObservabilityService, AssetService, CatalogService, ContractAgreementService, ContractDefinitionService, ContractNegotiationService, DataplaneSelectorService, HTTPProvisionerWebhookService, PolicyService, TransferProcessService];
+export const APIS = [AssetService, ContractAgreementService, ContractDefinitionService, ContractNegotiationService, PolicyService, TransferProcessService];
