@@ -3,7 +3,7 @@
  * EDC REST APIs - merged by OpenApiMerger
  *
  * The version of the OpenAPI document: 0.0.1-SNAPSHOT
- * 
+ *
  *
  * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
  * https://openapi-generator.tech
@@ -12,8 +12,7 @@
 import { ContractOffer } from './contractOffer';
 
 
-export interface Catalog { 
+export interface Catalog {
     contractOffers?: Array<ContractOffer>;
     id?: string;
 }
-
