@@ -3,17 +3,17 @@
  * EDC REST APIs - merged by OpenApiMerger
  *
  * The version of the OpenAPI document: 0.0.1-SNAPSHOT
- * 
+ *
  *
  * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
  * https://openapi-generator.tech
  * Do not edit the class manually.
  */
-import { Policy } from './policy';
-import { Asset } from './asset';
+import { Policy } from '.';
+import { Asset } from '.';
 
 
-export interface ContractOffer { 
+export interface ContractOffer {
     asset?: Asset;
     consumer?: string;
     contractEnd?: string;
